Stabilize record order of -fd2p and -fp2d
These functions push records into a dictionary that then repopulates a new seqbuddy object. This shuffles the order of the records though, which is a pain. Now the original order of the returned file is retained, as is the number of sequences (originally, records were deleted if not present in both files)