New function hash_ids()
Essentially the same function as the one in SeqBuddy. It actually calls the SeqBuddy function, instead of reimplementing the entire function.
Originally I considered using the same hash for sequences with the same ID in multiple alignments (multi-alignment files), but this breaks really bad when there are sequences with the same ID in a single alignment... Everyone gets their own hash.
Unit tests written and passing