Issue 49
This begins to address issue 49 (Reorganize repo to be more in line with
common Python practice) and adds a couple of initial pieces which will
be needed for registering the package on PyPI.

The workshop directory has been renamed to buddysuite.

LICENSE has been renamed to LICENSE.txt

A .pypirc file has been added which contains the URLs for
the pypi and pypitest repos.

A setup.cfg file has been added which simply points at README.md
as the source of metadata.

No functional changes to the code have been made. The install and uninstall
processes still work exactly as before with the rename of workshop to buddysuite.