Issue 47
The install process has been changed to use Distutils/setuptools
to install commandline scripts instead of editing a user's
.bashrc/.*profile, creating a hidden directory inside of their
install directory, and symlinking to the Python scripts.

As such, BuddySuite_installer.py (which was actually a binary) and
the Install_Package directory (which contained earlier stubs for
using Distutils/setuptools) have been removed.

A setup.py file and a config.ini file have been added.

The config.ini file is blank now, but will eventually contain 'flags'
to hold the user's choice for opting in to the SIP.

The setup.py file uses setuptools to install the buddysuite package
and create 'entry_points' for the *Buddy.py scripts. It also contains
information that will be useful when the package is registered on PyPI.

The install process is now in line with common Python practice. The user
can download the source and run 'python setup.py install' to install
the buddysuite package (into /usr/lib/python3.x/site-packages) and create
alignbuddy, databasebuddy, phylobuddy,and seqbuddy executables (which
are placed in /usr/bin).

The *Buddy.py files had to be modified in order to use setuptools'
'entry_points' feature in the following manner:

    if __name__ == '__main__':
        <body of main function>

was changed to

    def main():
        <body of main function>

    if __name__ == '__main__':
        main()

The next step will be to 'finalize' the CLASSIFIERS, KEYWORDS, etc.
data in setup.py, then test registering the package on PyPI. Instructions
for that process can be found here:

    http://peterdowns.com/posts/first-time-with-pypi.html