Remove the creation of a Conda virtual env in Docker, and swap update order
There was an issue with conda in the Docker container relating back to when the bioconda channel was added compared to when conda is updated.
By switching the channel before updating, the issue appears to be resolved.
I've also removed the conda virtual environment because it's unnecessary. It's just adding another level of complexity without adding any value.