Clean up do_load handling of temp files
The loaded http_error_file and results_file were being reset to files in a TempDir, which is a hold over from a previous implementation. I assume this was causing weird behaviour... Switch to TempFile() objects.
Also stop looping back into do_load if an incorrect path or non-db session file is pointed to. Just throw the error message and let the user type it in again.
New batch of tests included